requirements.txt
# YouTube Orchestrator MCP - Python Dependencies

# Core dependencies
requests>=2.31.0
pydantic>=2.0.0
filelock>=3.12.0
python-dotenv>=1.0.0

# YouTube API
youtube-transcript-api>=0.6.1

# CLI & utilities
tqdm>=4.66.0
python-slugify>=8.0.1
python-dateutil>=2.8.2
isodate>=0.6.1

# LLM & LangChain
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5

# MCP server
mcp>=0.9.0

# Optional (for future HTTP mode)
fastapi>=0.109.0
uvicorn>=0.27.0

